subscriber/requirements.txt
paho-mqtt==2.1.0
elasticsearch==9.0.0
dotenv==0.9.9